\name{bbObs}
\docType{data}
\alias{bbObs}
\title{Phenological observation: budburst}
\description{
This dataset gives an example of budburst observations from the German Weather Service (DWD) 
obtainable by the plant phenological online database PPODB. 
The dataset was extracted out of the downloadable tsv-File by use of \code{\link{data.read.phase}}.
}
\usage{bbObs}
\format{A dataframe containing three observations.}
\source{Plant Phenological Online Database (PPODB), \url{www.phenology.de}}
\references{Dierenbach, J., Badeck, F.W., Schaber, J., 2013. The plant phenological online database (PPODB): an online database for long-term phenological data. International Journal of Biometeorology , 1-8.}
\seealso{\code{\link{data.read.phase}},\code{\link{lcObs}}}
\keyword{datasets}
