\name{dataFinal}
\docType{data}
\alias{dataFinal}
\title{Preprocessed data}
\description{
This dataset contains phenological data (connected leafcolouring and budburst observations) as well as 
spatially related temperature data. The dataset was created out of \code{\link{extractedObs}}, \code{\link{relatedGrid}} and 
\code{\link{temperatures}} by use of \code{\link{data.addTemperatures}}.
}
\usage{data(dataFinal)}
\format{A dataframe containing three rows of connected phenological information combined with temperature data.}
\seealso{\code{\link{data.addTemperatures}},\code{\link{bbObs}},\code{\link{lcObs}},\code{\link{extractedObs}},\code{\link{relatedGrid}},\code{\link{temperatures}}}
\keyword{datasets}

