\name{extractedObs}
\docType{data}
\alias{extractedObs}
\title{Phenological observation: leafcolouring and budburst}
\description{
This dataset gives an example of leafcolouring observations connected with subsequent budburst observations. 
The observations are extracted out of phenological data from the German Weather Service (DWD) obtainable by 
the plant phenological online database PPODB.
The dataset was created out of the datasets \code{\link{bbObs}} and \code{\link{lcObs}} by use of method \code{\link{data.extract}}.
}
\usage{extractedObs}
\format{A dataframe containing three rows of connected observations.}
\source{Plant Phenological Online Database (PPODB), \url{www.phenology.de}}
\references{Dierenbach, J., Badeck, F.W., Schaber, J., 2013. The plant phenological online database (PPODB): an online database for long-term phenological data. International Journal of Biometeorology , 1-8.}
\seealso{\code{\link{data.extract}},\code{\link{bbObs}},\code{\link{lcObs}}}
\keyword{datasets}
