\name{relatedGrid}
\docType{data}
\alias{relatedGrid}
\title{Grid containing spatial information related to temperature data}
\description{
This dataset, based on a digital elevation model contains spatial information related to the temperature data.
}
\usage{relatedGrid}
\format{A dataframe containing spatial information.}
\source{digital elevation model}
\seealso{\code{\link{temperatures}}}
\keyword{datasets}
